Moving elaborate_only to test suites.
This is to prepare for re-factoring away the need for simulator
interfaces to monitory internal signals of the runner to end
simulation and determine sucess. In the future the simulator exit code
is ignored in favor of reading a file with test case results. However
when elaborating the exit code should not be ignored thus requiring
this re-factoring.