Always use the most up-to-date version of modelsim.ini
The local copy of modelsim.ini is now re-created on each invocation
of VUnit, to capture any potential updates to VUNIT_MODELSIM_INI.
The unit tests for ModelSimInterface have been updated to eliminate
the expectation that a local modelsim.ini file exits. Instead,
setUp includes the creation of a simulated modelsim environment.
Upon initialization, ModelSimInterface will pull the modelsim.ini
file from this environment, as it would in the real case.