Modified merge_coverage to handle long vcover merge commands.
The vcover command (launched with python subprocess) may exceed the
limit of 32768 characters in the Win32 API when merging too many
coverage files.

This is avoided by creating a temporary text file with the path to
all coverage files to be merged and applying the new text file using
the vcover merge -inputs flag.