Update ptr packages to support new queue_t implementation
Update the integer_vector_ptr and string_ptr packages to support the
changes necessary for a new implementation of queue_t.

both packages:
- Implement a stack to reuse the ref indices of deallocated pointers
- Add vec_t and va_t aliases
- Improve consistency in format and substance between the two packages

integer_vector_ptr_pkg:
- Change the drop parameter in resize to a rotate parameter. This
supports reallocating a circular buffer. This parameter is only
necessary to support the implementation of queue_t, so there is no need
to implement the functionality for string_ptr as well.

string_ptr_pkg:
- Remove support for the drop parameter when resizing a string_ptr. The
only use of it in the entire code base was to support the old
implementation of queue_t.