Update implementation of queue_t
Previously, queue_t was implemented as a single-string, linear buffer.
Update queue_t to use a circular buffer of integers, where each integer
represents a string_ptr. Each item contained within the queue is
represented by a string_ptr, where the first character is a type mark
and the remaining characters are the encoded form of the item.