+: add function get_minimal_file_set_in_compile_order
These function takes a list of target source files and returns the
minimal list of SourceFiles to compile for the given target source
files.
One quirk is that at the end of the function the list has to be
filtered. This has to do with the get_files_in_compile_order function,
which for reasons that are unknown to me introduces unnecessary files.