Drive invalid in AXI VCs
When the 'valid' signal is low for an AXI (-Lite) channel, the signals of the channel should be driven with 'X'.
They should not keep their value from the last time the channel was valid.
This can hide errors in the DUT if the DUT samples values in the wrong clock cycle.

This change is largely inspired by the AXI-Stream VCs.

Note that this changes the default behavior of the three VCs, so in some sense it is a breaking change.
This could be avoided by setting the default value 'drive_invalid' generic to 'false'.
Though having it set to 'true' is symmetric with the AXI-Stream VCs.
And in my personal opinion it should be enabled in pretty much all cases, to get a stronger test.