Adjust disable_ieee_warnings for NVC >= 1.16
For the next release of NVC this should be passed as a global option
rather than a "run" option (because warnings can be produced during
elaboration too).  Passing to the run command is still supported so old
versions of VUnit will continue to work, but the option may not take
effect in some situations.