Add tkeep and tstrb support for non-byte tdata widths in AXI stream VCs.
The AXI standard is limited to byte-oriented data buses and specifies the width of tkeep and tstrb
as tdata'length / 8. To support the use of non-byte tdata widths with active use
of tkeep and tstrb, this has been changed to (tdata'length + 7) / 8  such that all bits in tdata are
covered by a bit in tkeep and tstrb.